Moved query for missing item fields when crossposting to use-crossposter hook
